Vakitacontroller

See merge request ctd/hispanos/proyecto-integrador-2/proyecto-integrador-0723/1021ft-c1/equipo-03!7

+++ api-vakita/ms-vakitas/src/main/java/com/grupo3/msvakitas/service/impl/VakitaService.java
@@ -103,7 +103,7 @@ public class VakitaService implements IVakitaService {
             Double deposit = vakitaModify.getCumulativeAmount() + amount;
             vakitaModify.setCumulativeAmount(deposit);
             log.info("Success, amount update: "+ vakitaModify.getCumulativeAmount());
-            this.updateVakita(vakitaModify);
+            this.updateVakita(id, vakitaModify);
         }
         else if(!vakitaModify.getIsActive()){
             throw new BadRequestException("Para depositar dinero la vakita debe estar activa");
@@ -155,16 +155,21 @@ public class VakitaService implements IVakitaService {
     }
 
     //Este método es para actualizar una vakita existente
+    //LOS CAMPOS QUE SE PERMITEN MODIFICAR SON: Descripcion, imagen, fecha de expiracion
     @Override
-    public VakitaDTO updateVakita(VakitaDTO vakita) throws BadRequestException {
+    public VakitaDTO updateVakita(Long id, VakitaDTO vakita) throws BadRequestException, ResourceNotFoundException {
+         VakitaDTO vakitaToModify = this.getVakitaById(id);
         if (vakita.getExpirationDate().equals(LocalDate.now()) || vakita.getName() == null || vakita.getIdCreatorUser() == null){
             throw new BadRequestException("Alguno de los datos es incorrecto, no se puede actualizar la vakita");
         }
-        else if(vakitaRepository.existsById(vakita.getId())){
-            Vakita vakitaToSave = mapper.map(vakita, Vakita.class);
+        vakitaToModify.setDescription(vakita.getDescription());
+        vakitaToModify.setImgURL(vakita.getImgURL());
+        vakitaToModify.setExpirationDate(vakita.getExpirationDate());
+        if(vakitaRepository.existsById(vakitaToModify.getId())){
+            Vakita vakitaToSave = mapper.map(vakitaToModify, Vakita.class);
             vakitaRepository.save(vakitaToSave);
         }
-        log.info("Update vakita id: " + vakita.getId());
+        log.info("Update vakita id: " + id);
         return vakita;
     }
 
@@ -174,18 +179,22 @@ public class VakitaService implements IVakitaService {
         VakitaDTO vakita = this.getVakitaById(vakitaId);
         List<UserDTO> listContributors =  vakita.getContributors();
         listContributors.add(user);
-        this.updateVakita(vakita);
+        this.updateVakita(vakitaId, vakita);
         log.info("Success updating vakita: " + vakitaId);
     }
 
     //Método para eliminar una vakita
     @Override
-    public void deleteVakita(Long id) throws ResourceNotFoundException{
+    public void deleteVakita(Long id) throws ResourceNotFoundException, BadRequestException {
         VakitaDTO vakitaToDrop = this.getVakitaById(id);
-        Vakita vakitaEntityToDrop = mapper.map(vakitaToDrop, Vakita.class);
-        vakitaRepository.delete(vakitaEntityToDrop);
-        log.info("Success, vakita deleted: " + id);
-
+        if (vakitaToDrop.getIsActive() == true ){
+            throw new BadRequestException("Para eliminar un vakita debe estar inactiva");
+        }
+        else {
+            Vakita vakitaEntityToDrop = mapper.map(vakitaToDrop, Vakita.class);
+            vakitaRepository.delete(vakitaEntityToDrop);
+            log.info("Success, vakita deleted: " + id);
+        }
     }
 
     //Este método es para dejar una vakita inactiva, es decir
@@ -198,7 +207,7 @@ public class VakitaService implements IVakitaService {
             vakitaToCancel.setIsActive(false);
         }
         log.info("Success, cancel vakita id: " + id);
-        this.updateVakita(vakitaToCancel);
+        this.updateVakita(id, vakitaToCancel);
     }
 
 }
